% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cncalc.R
\name{cncalc}
\alias{cncalc}
\title{This function is utilized to determine the optimal number of clusters for k-means clustering on multi-sample omics data.}
\usage{
cncalc(data)
}
\arguments{
\item{data}{A time-series omics matrix.}
}
\description{
This function is utilized to determine the optimal number of clusters for k-means clustering on multi-sample omics data.
}
