% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansR.R, R/KmeansR2.R
\name{KmeansR}
\alias{KmeansR}
\alias{KmeansR2}
\title{This function is designed for computing k-means clustering and creating visualizations for multi-sample omics data.}
\usage{
KmeansR(data, centers, table = FALSE, angle = 90, box = FALSE)

KmeansR2(
  data,
  centers = 6,
  table = TRUE,
  angle = 90,
  box = FALSE,
  label_size = 3,
  legend.position = "right"
)
}
\arguments{
\item{data}{A time-series omics matrix.}

\item{centers}{Number of cluster.}

\item{table}{Logical value,If TRUE is selected, a kmeans_result.csv will be output.The content involves performing z-score normalization on each row of the data frame, along with information about different clusters.}

\item{angle}{The angle of rotation for x-axis labels.}

\item{box}{logical. Should a border be drawn around the plot?}

\item{label_size}{Label Size on Bar Chart}

\item{legend.position}{Legend Position on Bar Chart. Specify the desired location of legends, choosing from options such as "none," "left," "right," "bottom," "top," or a two-element numeric vector.}
}
\description{
This function is designed for computing k-means clustering and creating visualizations for multi-sample omics data.

This function is designed for computing k-means clustering and creating visualizations for multi-sample omics data.
}
\examples{
library(KmeansTrendAnalyzer)
library(tidyverse)
data(mango_FI_DAM)
df <- mango_FI_DAM \%>\%
  select(Index,`TA-1`:`C-3`) \%>\%
  column_to_rownames("Index") \%>\%
  mutate(TA = rowMeans(select(., 1:3)),
         TB = rowMeans(select(., 4:6)),
         TC = rowMeans(select(., 7:9)),
         A = rowMeans(select(., 10:12)),
         B = rowMeans(select(., 13:15)),
         C = rowMeans(select(., 16:18))) \%>\%
  select(!contains("-"))
cncalc(df)
KmeansR(df,centers = 3,angle = 0,box = TRUE)
library(KmeansTrendAnalyzer)
library(tidyverse)
data <- mango_FI_DAM \%>\%
  select(Index,`TA-1`:`C-3`,Class = Class.I) \%>\%
  column_to_rownames("Index") \%>\%
  mutate(TA = rowMeans(select(., 1:3)),
         TB = rowMeans(select(., 4:6)),
         TC = rowMeans(select(., 7:9)),
         A = rowMeans(select(., 10:12)),
         B = rowMeans(select(., 13:15)),
         C = rowMeans(select(., 16:18))) \%>\%
  select(!contains("-")) \%>\%
  mutate(Colour = case_when(
    Class == "Amino acids and derivatives" ~ "#FF0000",
    Class == "Lipids" ~ "#FFFF00",
    Class == "Nucleotides and derivatives" ~ "#00FF00",
    Class == "Organic acids" ~ "#00FFFF",
    Class == "Phenolic acids" ~ "#0000FF",
    Class == "Others" ~ "#FF00FF"
  ))
df_mango <- data
KmeansR2(df_mango)
# ggsave("df_mango.pdf",width = 20,height = 10)
}
